Add basic structure for scss file and React components
