+++ package.json
@@ -8,6 +8,7 @@
 		"@testing-library/user-event": "^13.5.0",
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0",
+		"react-router-dom": "^6.22.0",
 		"react-scripts": "5.0.1",
 		"web-vitals": "^2.1.4"
 	},
@@ -36,11 +37,16 @@
 		]
 	},
 	"devDependencies": {
+		"@types/jest": "^29.5.12",
+		"@types/node": "^20.11.16",
+		"@types/react": "^18.2.52",
+		"@types/react-dom": "^18.2.18",
 		"eslint": "^8.56.0",
 		"eslint-config-airbnb": "^19.0.4",
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-prettier": "^5.1.3",
 		"prettier": "^3.2.5",
-		"sass": "^1.70.0"
+		"sass": "^1.70.0",
+		"typescript": "^5.3.3"
 	}
 }
