--- src/App.jsx
@@ -1,15 +1,0 @@
-import './App.css';
-import React from 'react';
-
-function App() {
-	return (
-		<div className="App">
-			<header className="App-header">
-				<h1>House Booking App</h1>
-				<p>coming soon</p>
-			</header>
-		</div>
-	);
-}
-
-export default App;
