+++ src/App.tsx
@@ -1,0 +1,13 @@
+import './styles/index.scss';
+import React from 'react';
+import LandingPage from './views/landingpage/LandingPage';
+
+function App() {
+	return (
+		<div className="App">
+			<LandingPage />
+		</div>
+	);
+}
+
+export default App;
