+++ src/components/Footer.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Footer() {
+	return <footer>This is a footer</footer>;
+}
+
+export default Footer;
