+++ src/components/Header.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+function Header() {
+	return (
+		<header>
+			<nav>navigation</nav>
+			<div className="hero">
+				<h1 className="hero">House booking app</h1>
+				<p className="summary">coming soon</p>
+			</div>
+		</header>
+	);
+}
+
+export default Header;
