+++ src/styles/index.scss
@@ -1,0 +1,4 @@
+@import './base/fonts.scss';
+@import './base/mixins.scss';
+@import './base/typography.scss';
+@import './base/variables.scss';
