+++ src/views/landingpage/LandingPage.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import Footer from '../../components/Footer';
+import Header from '../../components/Header';
+
+function LandingPage() {
+	return (
+		<>
+			<Header />
+			<main>
+				<div className="calendar-wrapper">Calendar</div>
+			</main>
+			<Footer />
+		</>
+	);
+}
+
+export default LandingPage;
